builder-worker: support multiple integration types in runner::util
Until now the validation and generating the spec assumed only one
element, docker integration. This commit filters both for integration
type.

The validation function was split to allow handling the checks per
integration type.

+++ components/builder-worker/src/runner/kubernetes.rs
@@ -36,7 +36,7 @@ const KUBECONFIG_ENVVAR: &'static str = "KUBECONFIG";
 
 pub struct KubernetesExporterSpec {
     pub kubeconfig_path: String,
-    pub replicas: i32,
+    pub replicas: i64,
 }
 
 pub struct KubernetesExporter<'a> {
